coint4/pyproject.toml
[tool.poetry]
name = "coint2"
version = "0.1.1"
description = "Cointegration pairs trading framework for cryptocurrencies"
authors = ["Trading Team <team@coint2.io>"]
license = "MIT"
readme = "README.md"
packages = [{include = "coint2", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "*"
joblib = "*"
pydantic = "*"
pyyaml = "^6.0"
numpy = "*"
pandas = "*"
scipy = "^1.13.0"
numba = "*"
# ↓↓↓ ЭТИ ЗАВИСИМОСТИ МЫ БЕРЕМ ИЗ ВЕТКИ CODEX ↓↓↓
pyarrow = ">=8.0.0"
statsmodels = "*"
dask = {extras = ["dataframe"], version = "*"} # Устанавливаем dask с поддержкой DataFrame
dask-expr = "*" # Нужен для read_parquet в новых версиях
python-dotenv = "*"
matplotlib = "*"  # Для графиков
seaborn = "*"     # Для красивых графиков
plotly = "*"      # Для интерактивных графиков
polars = "^1.31.0"
tqdm = "^4.67.1"
hurst = "*"
psutil = "*"
scikit-learn = "*"
tabulate = "^0.9.0"  # Для красивого вывода таблиц

# Moved to tool.poetry.group.dev.dependencies
optuna = "^4.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4"
mypy = "^1.10"
pytest = "*"
pytest-xdist = "*"  # Для параллельного запуска тестов (-n auto)
pytest-timeout = "*"  # Для таймаутов на тесты
pytest-rerunfailures = "*"  # Для повторного запуска флаки тестов
pytest-cov = "*"  # Для покрытия кода тестами
filelock = "*"  # Для кроссплатформенной блокировки файлов

[tool.poetry.scripts]
coint2 = "coint2.cli:main"
coint2-optimize = "coint2.optimiser.run_optimization:cli_main"
coint2-live = "coint2.cli_live:main"
coint2-check-health = "coint2.cli.check_coint_health:main"
coint2-build-universe = "coint2.cli.build_universe:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
select = ["E", "F", "B", "I", "UP"]

